% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{calc_peak_metrics}
\alias{calc_peak_metrics}
\title{calc_peak_metrics}
\usage{
calc_peak_metrics(storage_object)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}
}
\value{
an updated metScribeR storage_object
}
\description{
Calculates quality metrics and other feature-level information.
}
\details{
Calculates the following: feature retention time at apex, feature intensity at apex, feature standard deviation,
feature skew, feature upper and lower rt boundaries, feature width at half max, total number of eic points in a feature,
the number of features found in a feature's adduct, and quality metrics beta_shape_parameter, novel_snr, and correlation with beta curve
from Kumler, W., Hazelton, B.J. & Ingalls,  A.E. Picky with peakpicking: assessing chromatographic peak
quality with simple metrics in metabolomics. BMC Bioinformatics 24, 404 (2023). https://doi.org/10.1186/s12859-023-05533-4.
}
