% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eic.R
\name{create_eic}
\alias{create_eic}
\title{create_eic}
\usage{
create_eic(data, mz_list, eic_mz_tolerance)
}
\arguments{
\item{data}{the data from parse_mzML function output}

\item{mz_list}{a numeric vector of the mz values to search for}

\item{eic_mz_tolerance}{the mz tolerance for eic creation in ppm}
}
\value{
an eic as a tibble
}
\description{
Create an eic given data, a list of mzs to look for, and a mz tolerance
}
