% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{eic_plot}
\alias{eic_plot}
\title{eic_plot}
\usage{
eic_plot(storage_object, unique_standard_id_plotting)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}

\item{unique_standard_id_plotting}{the unique_standard_id for the standard being plotted. This will be an integer.}
}
\value{
a ggplot object
}
\description{
plots an eic figure for the given unique_standard_id
}
\details{
For a given standard, plots a line graph of all adducts, colored by identity.
}
