% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{export_library_csv}
\alias{export_library_csv}
\title{export_library_csv}
\usage{
export_library_csv(storage_object, save_file_path)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}

\item{save_file_path}{the location the csv file will be saved to}
}
\description{
Exports a library csv file
}
