% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{export_library_dataframe}
\alias{export_library_dataframe}
\title{export_library_dataframe}
\usage{
export_library_dataframe(storage_object)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}
}
\value{
a dataframe object for display in metScribeR
}
\description{
Exports a library dataframe
}
