% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{find_adduct_conflicts}
\alias{find_adduct_conflicts}
\title{find_adduct_conflicts}
\usage{
find_adduct_conflicts(
  storage_object,
  mz_tolerance = storage_object$library_data$eic_mz_tolerance,
  rt_tolerance
)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}

\item{mz_tolerance}{the tolerance for a conflict, in ppm. For example mz_tolerance = 30 would be 30 ppm.}

\item{rt_tolerance}{the tolerance for a conflict, in min.}
}
\value{
an updated metScribeR storage_object
}
\description{
Find conflicting adducts in the library
}
\details{
This function finds features that overlap with a adduct peak in the library within an RT and m/z tolerance. It reports
these conflicts in the adduct_df with the name, adduct, and rt of the conflicting feature. Note that this looks for
conflicting features, not necesarilly conflicting confirmed identified adduct species.
}
