% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{find_eic_peaks}
\alias{find_eic_peaks}
\title{find_eic_peaks}
\usage{
find_eic_peaks(storage_object)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}
}
\value{
an updated metScribeR storage_object
}
\description{
Feature finding for eics
}
\details{
This function is primarily a wrapper around chromatographR::find_peaks() with the
addition of an additional density filtering step that removes outlier points.
It loops through all the eics in the active library and puts the resulting peaks in
the peaks_df slot of the metScriber storage_object.
}
