% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{import_eic_data}
\alias{import_eic_data}
\title{import_eic_data}
\usage{
import_eic_data(storage_object)
}
\arguments{
\item{storage_object}{a metScribeR object}
}
\value{
an updated metScribeR object
}
\description{
import_eic_data
}
\details{
Imports mzML data and creates EICs for each adduct*standard combination provided with the adduct_search_df and standard_df
These are stored in the storage_object for use with further functions. Runs in parallel and could take several hours to run
depending on the number of files being processed. Only functions on the active library.
}
