% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{manual_peak_annotation}
\alias{manual_peak_annotation}
\title{manual_peak_annotation}
\usage{
manual_peak_annotation(storage_object, unique_peak_id_annotate, annotation)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}

\item{unique_peak_id_annotate}{the unique_peak_id of the feature to recieve annotation}

\item{annotation}{logical description of peak quality. T if good, F if bad}
}
\value{
a metScribeR storage_object
}
\description{
Add manual quality assessment result to a peak
}
