% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{noise_plot}
\alias{noise_plot}
\title{noise_plot}
\usage{
noise_plot(plotting_mzML_file_path, noise_level_plotting = NA)
}
\arguments{
\item{plotting_mzML_file_path}{path to file for plotting}

\item{noise_level_plotting}{optional argument to add a horizontal line at the indicated intensity value}
}
\value{
a ggplot object
}
\description{
plots an spectra figure for the given mzML file, indicated with a unique_standard_id and a mode.
}
\details{
This plot is useful for choosing the background noise level while importing data.
}
