% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mzML.R
\name{parse_mzML}
\alias{parse_mzML}
\title{parse_mzML}
\usage{
parse_mzML(path, noise_level, mode = NA)
}
\arguments{
\item{path}{path to the mzML file to be read}

\item{noise_level}{level below which all MS observations will be excluded.}

\item{mode}{ESI mode of data, "POS" or "NEG"

@export}
}
\description{
This function returns data in an R dataframe from mzML format
}
\details{
This function is a wrapper around mzR::openMSfile(), mzR::header(), & mzR::peaks(). It is
similar to what Spectra does, but faster in the metScribeR context. Designed for use only for MS1 data.
}
