% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{peak_demo_plot}
\alias{peak_demo_plot}
\title{peak_demo_plot}
\usage{
peak_demo_plot(storage_object, unique_adduct_id_plotting)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}

\item{unique_adduct_id_plotting}{the unique_adduct_id for the feature being plotted. This will be an integer.}
}
\value{
a ggplot object
}
\description{
plots an eic figure for the given unique_adduct_id. This plot is specific to the find peaks demo, and is not for the review peaks section.
}
