% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{peak_plot}
\alias{peak_plot}
\title{peak_plot}
\usage{
peak_plot(storage_object, unique_peak_id_plotting)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}

\item{unique_peak_id_plotting}{the unique_peak_id for the feature being plotted. This will be an integer.}
}
\value{
a ggplot object
}
\description{
plots an eic figure for the given unique_peak_id. This is just one feature, in contrast to
one standard as with eic_plot
}
\details{
For a given feature, plots a line graph of eic points.
}
