% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{peak_review_plot}
\alias{peak_review_plot}
\title{peak_review_plot}
\usage{
peak_review_plot(storage_object, unique_peak_id_plotting)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}

\item{unique_peak_id_plotting}{the unique_peak_id for the feature being plotted. This will be an integer.}
}
\value{
a ggplot object
}
\description{
plots an eic figure for the given unique_peak_id. This plot is specific to the review peaks tab.
}
