% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{save_storage_object}
\alias{save_storage_object}
\title{save_storage_object}
\usage{
save_storage_object(storage_object, save_file_path)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}

\item{save_file_path}{the location the rds file will be saved to}
}
\description{
Exports a rds file of the storage_object.
}
