% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object.R
\name{storage_object}
\alias{storage_object}
\title{storage_object}
\usage{
storage_object()
}
\description{
initializes a metScribeR storage_object
}
\details{
Calling this function creates a blank storage_object, including library_data, standard_df,
adduct_df, peak_df, and data_df.
}
