% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{update_best_peak_and_adduct}
\alias{update_best_peak_and_adduct}
\title{update_best_peak_and_adduct}
\usage{
update_best_peak_and_adduct(
  storage_object,
  apex_intensity_ratio_tolerance = 0.5
)
}
\arguments{
\item{storage_object}{a metScribeR storage_object}

\item{apex_intensity_ratio_tolerance}{the intensity tolerance for a peak to be considered multi-modal within an EIC}
}
\value{
an updated metScribeR storage_object
}
\description{
Adds the unique peak id of the best peak to adduct_df. Adds the is_best_adduct label to adduct_df.
}
\details{
The best peak for an adduct is the one with the maximum intensity at its apex. This is because we are
working with authentic standard data, where we assume the standard will be more intense than any noise.

The best adduct is one which has a peak that has passed manual quality classification
and has the greatest apex intensity. The representative peak for an adduct is determined by the rules in the
prior paragraph. This only updates after all adducts have been evaluated for a standard.
}
