% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_object_functions.R
\name{user_data_input}
\alias{user_data_input}
\title{user_data_input}
\usage{
user_data_input(
  storage_object,
  standard_df,
  adduct_search_df,
  eic_mz_tolerance,
  eic_rt_tolerance,
  mzML_parsing_noise_level,
  output_directory_path
)
}
\arguments{
\item{storage_object}{A metScribeR object}

\item{standard_df}{An R dataframe with the following columns:
common_name as character, monoisotopic_mass as numeric, pos_mode_mzML_file_path
as character, neg_mode_mzML_file_path as character. File paths should be absolute,
beginning with C:/ or H:/ or similar.

This dataframe is the input related to the standards being processed by metScribeR.}

\item{adduct_search_df}{An R dataframe with the following columns:
adduct as character, change_from_neutral as numeric, and mode as character.
For example: data.frame(adduct = 'M+H', change_from_neutral = 1.008 amu, mode = 'POS')

This dataframe is the input related to which adducts will be searched for each standard.}

\item{eic_mz_tolerance}{a numeric value giving the m/z window for eic creation in PPM.}

\item{eic_rt_tolerance}{a numeric value giving the rt window distinguishing adducts in min.}

\item{mzML_parsing_noise_level}{Describes the level of intensity in the raw
data below which all signals are discarded as noise.}
}
\value{
an updated storage_object
}
\description{
Collects data and parameters from user and adds them to the storage_object.
}
\details{
The data and parameters collected here are for eic creation. Data are stored
in the appropriate table/slot in the storage_object where they can be used in later computation.

These parameters are associated with the active library id.
}
